Modified SC with employee
