--- chaincode/artemis/uhcl_international1/MyAsset/index.js
@@ -1,5 +1,0 @@
-const { MyAsset } = require('./lib/MyAsset')
-
-module.exports = {
-    MyAsset
-}
