+++ tmpworkspaceforuser/EmployeeAsset/index.js
@@ -1,0 +1,5 @@
+const { EmployeeAsset } = require('./lib/EmployeeAsset')
+
+module.exports = {
+    EmployeeAsset
+}
