--- tmpworkspaceforuser/MyAsset/index.js
@@ -1,5 +1,0 @@
-const { MyAsset } = require('./lib/MyAsset')
-
-module.exports = {
-    MyAsset
-}
